Split markdown-renderer.tsx
