+++ src/components/editor/scroll/utils.ts
@@ -1,4 +1,5 @@
-import { LineMarkerPosition } from '../../markdown-renderer/markdown-renderer'
+import { LineMarkerPosition } from '../../markdown-renderer/basic-markdown-renderer'
+
 export const findLineMarks = (lineMarks: LineMarkerPosition[], lineNumber: number): { lastMarkBefore: LineMarkerPosition | undefined, firstMarkAfter: LineMarkerPosition | undefined } => {
   let lastMarkBefore
   let firstMarkAfter
