+++ src/components/markdown-renderer/utils/html-react-transformer.tsx
@@ -1,49 +1,15 @@
-import { diffArrays } from 'diff'
 import { DomElement } from 'domhandler'
 import React, { Fragment, ReactElement } from 'react'
 import { convertNodeToElement, Transform } from 'react-html-parser'
 import {
   ComponentReplacer,
   SubNodeTransform
-} from './replace-components/ComponentReplacer'
+} from '../replace-components/ComponentReplacer'
+import { LineKeys } from './types'
 
 export interface TextDifferenceResult {
   lines: LineKeys[],
   lastUsedLineId: number
-}
-
-export interface LineKeys {
-  line: string,
-  id: number
-}
-
-export const calculateNewLineNumberMapping = (newMarkdownLines: string[], oldLineKeys: LineKeys[], lastUsedLineId: number): TextDifferenceResult => {
-  const lineDifferences = diffArrays<string, LineKeys>(newMarkdownLines, oldLineKeys, {
-    comparator: (left:string|LineKeys, right:string|LineKeys) => {
-      const leftLine = (left as LineKeys).line ?? (left as string)
-      const rightLine = (right as LineKeys).line ?? (right as string)
-      return leftLine === rightLine
-    }
-  })
-
-  const newLines: LineKeys[] = []
-
-  lineDifferences
-    .filter((change) => change.added === undefined || !change.added)
-    .forEach((value) => {
-      if (value.removed) {
-        (value.value as string[])
-          .forEach(line => {
-            lastUsedLineId += 1
-            newLines.push({ line: line, id: lastUsedLineId })
-          })
-      } else {
-        (value.value as LineKeys[])
-          .forEach((line) => newLines.push(line))
-      }
-    })
-
-  return { lines: newLines, lastUsedLineId: lastUsedLineId }
 }
 
 export const calculateKeyFromLineMarker = (node: DomElement, lineKeys?: LineKeys[]): number|undefined => {
