+++ src/components/markdown-renderer/utils/line-number-mapping.ts
@@ -1,0 +1,32 @@
+import { diffArrays } from 'diff'
+import { TextDifferenceResult } from './html-react-transformer'
+import { LineKeys } from './types'
+
+export const calculateNewLineNumberMapping = (newMarkdownLines: string[], oldLineKeys: LineKeys[], lastUsedLineId: number): TextDifferenceResult => {
+  const lineDifferences = diffArrays<string, LineKeys>(newMarkdownLines, oldLineKeys, {
+    comparator: (left:string|LineKeys, right:string|LineKeys) => {
+      const leftLine = (left as LineKeys).line ?? (left as string)
+      const rightLine = (right as LineKeys).line ?? (right as string)
+      return leftLine === rightLine
+    }
+  })
+
+  const newLines: LineKeys[] = []
+
+  lineDifferences
+    .filter((change) => change.added === undefined || !change.added)
+    .forEach((value) => {
+      if (value.removed) {
+        (value.value as string[])
+          .forEach(line => {
+            lastUsedLineId += 1
+            newLines.push({ line: line, id: lastUsedLineId })
+          })
+      } else {
+        (value.value as LineKeys[])
+          .forEach((line) => newLines.push(line))
+      }
+    })
+
+  return { lines: newLines, lastUsedLineId: lastUsedLineId }
+}
