+++ src/components/markdown-renderer/utils/types.d.ts
@@ -1,0 +1,4 @@
+export interface LineKeys {
+  line: string,
+  id: number
+}
